update: check year & week

+++ f0016/f0016-a.sh
@@ -11,13 +11,22 @@ new_db_name=$7
 for file in $folder/"$prefix"*.gz; do
   filename=$(basename -- "$file")
   filename="${filename%.*}"
-  year="${filename:0:4}"
-  left="${filename#*_W}"
-  week="${left%_T*}"
-  echo $year
-  echo $week
-  mongorestore --gzip --archive=$file --nsFrom "${db_name}.*" --nsTo "${new_db_name}.*" \
-  && python3 $py_name $filename $new_db_name \
-  && echo $filename > geoname_latest_collection.txt \
-  && mongo $new_db_name --eval "db.getCollection('$filename').drop()"
+  year=${filename:0:4}
+  left=${filename#*_W}
+  week=${left%_T*}
+  if [[ $year -eq $latest_year ]]; then
+    if [[ $week -gt $latest_week ]]; then
+      mongorestore --gzip --archive=$file --nsFrom "${db_name}.*" --nsTo "${new_db_name}.*" \
+      && python3 $py_name $filename $new_db_name \
+      && echo $filename > geoname_latest_collection.txt \
+      && mongo $new_db_name --eval "db.getCollection('$filename').drop()"
+    fi
+  else
+    if [[ $year -gt $latest_year ]]; then
+      mongorestore --gzip --archive=$file --nsFrom "${db_name}.*" --nsTo "${new_db_name}.*" \
+      && python3 $py_name $filename $new_db_name \
+      && echo $filename > geoname_latest_collection.txt \
+      && mongo $new_db_name --eval "db.getCollection('$filename').drop()"
+    fi
+  fi
 done
